backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
supabase==2.22.0
python-dotenv==1.2.1
psycopg2-binary==2.9.7
gunicorn==21.2.0